--- frontend/src/components/Navbar.jsx
@@ -1,45 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-const Navbar = ({ user }) => {
-  return (
-    <>
-    <nav className="navbar navbar-expand-lg bg-body-tertiary">
-  <div className="container-fluid">
-    <Link className="navbar-brand" to="/">Navbar</Link>
-    <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-      <span className="navbar-toggler-icon"></span>
-    </button>
-    <div className="collapse navbar-collapse" id="navbarSupportedContent">
-      <ul className="navbar-nav me-auto mb-2 mb-lg-0">
-        {!user  && 
-        <>
-        <li className="nav-item">
-          <Link className="nav-link active" aria-current="page" to="/">Home</Link>
-        </li>
-        <li className="nav-item">
-          <Link className="nav-link active" aria-current="page" to="/signin">Signin</Link>
-        </li>
-        <li className="nav-item">
-          <Link className="nav-link active" aria-current="page" to="/signup">Signup</Link>
-        </li>
-        </>}
-        {user  && 
-        <>
-        <li className="nav-item">
-          <Link className="nav-link active" aria-current="page" to="#">{user.name}</Link>
-        </li>
-        <li className="nav-item">
-          <Link className="nav-link active" aria-current="page" to="/logout">Log out</Link>
-        </li>
-        </>}
-
-      </ul>
-    </div>
-  </div>
-</nav>
-    </>
-  )
-}
-
-export default Navbar
