+++ frontend/src/components/index.jsx
@@ -1,0 +1,2 @@
+export { default as Navbar } from "./Navbar/Navbar";
+export { default as Footer } from "./Footer/Footer";
