+++ frontend/src/navigation/authRoute.jsx
@@ -1,21 +1,18 @@
-import React from 'react'
-import Home from './../pages/Home';
-import Signin from './../pages/Signin';
-import Signup from './../pages/Signup';
-import {Routes, Route} from 'react-router-dom';
-import Error from './../pages/Error';
+import React from "react";
+import { Home, Signin, Signup, Error } from "./../pages/";
+import { Routes, Route } from "react-router-dom";
 
 const AuthRoute = () => {
   return (
     <>
-    <Routes>
-        <Route path='/' element={<Home />} />
+      <Routes>
+        <Route path="/" element={<Home />} />
         <Route path="*" element={<Error />} />
-        <Route path='/signup' element={<Signup />} />
-        <Route path='/signin' element={<Signin />} />
-    </Routes>
+        <Route path="/signup" element={<Signup />} />
+        <Route path="/signin" element={<Signin />} />
+      </Routes>
     </>
-  )
-}
+  );
+};
 
-export default AuthRoute
+export default AuthRoute;
