+++ frontend/src/navigation/userRoute.jsx
@@ -1,19 +1,19 @@
-import React from 'react';
-import {Routes, Route} from 'react-router-dom';
-import Logout from '../common/logout';
-import Dashboard from './../pages/Dashboard';
-import Error from './../pages/Error';
+import React from "react";
+import { Routes, Route } from "react-router-dom";
+import Logout from "../common/logout";
+
+import { Dashboard, Error } from "./../pages/";
 
 const UserRoute = () => {
   return (
     <>
-    <Routes>
+      <Routes>
         <Route path="*" element={<Error />} />
-        <Route path='/dashboard' element={<Dashboard />} />
-        <Route path='/logout' element={<Logout />} />
-    </Routes>
+        <Route path="/dashboard" element={<Dashboard />} />
+        <Route path="/logout" element={<Logout />} />
+      </Routes>
     </>
-  )
-}
+  );
+};
 
-export default UserRoute
+export default UserRoute;
