+++ frontend/src/pages/index.jsx
@@ -1,0 +1,6 @@
+export { default as Dashboard } from "./Dashboard/Dashboard";
+export { default as Error } from "./Error/Error";
+export { default as Home } from "./Home/Home";
+export { default as Signin } from "./Signin/Signin";
+export { default as Signup } from "./Signup/Signup";
+export { default as Welcome } from "./Welcome/Welcome";
